Handle case when item list is behind system fields (#57)
Keys in JSON aren't ordered, so far LDLite has been implicitly relying on the first key in the response to be the one that has the items. This has worked so far but starting in Sunflower keys seem a little weirder. Now we're iterating through all the known "not a list of records" keys when trying to locate the actual key.