Use the less restrictive datatype when a column is mixed (#59)
When scanning the data for column types, currently the "last" record wins and overwrites any other datatype that has been found. I'm not quite sure why it iterates over every record this way but I'm trying not to touch the existing transform code unless absolutely necessary.

This change checks the datatype this column was previously parsed as and then doesn't allow a type to be "upgraded" from a varchar to something more specific because that will cause issues inserting the actual data later.